requirements.txt
numpy
scikit-learn
pandas
xgboost
streamlit
Pillow